refactor: Util 의존을 기존 클래스로 분리
