+++ tomcat/src/main/java/nextstep/jwp/controller/DefaultHandler.java
@@ -5,9 +5,8 @@ import org.apache.coyote.model.request.HttpRequest;
 import org.apache.coyote.model.response.HttpResponse;
 import org.apache.coyote.model.response.ResponseLine;
 import org.apache.coyote.model.response.StatusCode;
-import org.apache.coyote.utils.Util;
 
-import static org.apache.coyote.utils.Util.getExtension;
+import static org.apache.coyote.model.request.RequestLine.getExtension;
 
 public class DefaultHandler extends AbstractHandler {
 
@@ -24,7 +23,7 @@ public class DefaultHandler extends AbstractHandler {
     public String getResponse(final HttpRequest httpRequest) {
         final String path = httpRequest.getPath();
         final String extension = ContentType.getType(getExtension(path));
-        final String body = Util.getResponseBody(path, this.getClass());
+        final String body = HttpResponse.getResponseBody(path, this.getClass());
 
         return HttpResponse.of(extension, body, ResponseLine.of(StatusCode.OK))
                 .getResponse();
