+++ tomcat/src/main/java/nextstep/jwp/controller/RegisterHandler.java
@@ -4,18 +4,19 @@ import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.coyote.model.request.HttpRequest;
 import org.apache.coyote.model.request.RequestBody;
+import org.apache.coyote.model.response.HttpResponse;
 import org.apache.coyote.model.response.StatusCode;
-import org.apache.coyote.utils.Util;
 
-import static org.apache.coyote.utils.Util.createResponse;
+import static org.apache.coyote.model.response.HttpResponse.createResponse;
+
 
 public class RegisterHandler extends AbstractHandler {
 
-    public static final String REGISTER_HTML = "/register.html";
-    public static final String INDEX_HTML = "/index.html";
-    public static final String ACCOUNT = "account";
-    public static final String PASSWORD = "password";
-    public static final String EMAIL = "email";
+    private static final String REGISTER_HTML = "/register.html";
+    private static final String INDEX_HTML = "/index.html";
+    private static final String ACCOUNT = "account";
+    private static final String PASSWORD = "password";
+    private static final String EMAIL = "email";
     private static final RegisterHandler INSTANCE = new RegisterHandler();
     private static final String CLIENT_ERROR_404 = "/404.html";
 
@@ -25,20 +26,20 @@ public class RegisterHandler extends AbstractHandler {
 
     @Override
     protected String getMethodResponse(final HttpRequest httpRequest) {
-        return createResponse(StatusCode.OK, Util.getResponseBody(REGISTER_HTML, getClass()))
+        return createResponse(StatusCode.OK, HttpResponse.getResponseBody(REGISTER_HTML, getClass()))
                 .getResponse();
     }
 
     @Override
     protected String postMethodResponse(final HttpRequest httpRequest) {
         saveUser(httpRequest);
-        return createResponse(StatusCode.FOUND, Util.getResponseBody(INDEX_HTML, getClass()))
+        return createResponse(StatusCode.FOUND, HttpResponse.getResponseBody(INDEX_HTML, getClass()))
                 .getResponse();
     }
 
     @Override
     protected String otherMethodResponse(final HttpRequest httpRequest) {
-        return createResponse(StatusCode.UNAUTHORIZED, Util.getResponseBody(CLIENT_ERROR_404, getClass()))
+        return createResponse(StatusCode.UNAUTHORIZED, HttpResponse.getResponseBody(CLIENT_ERROR_404, getClass()))
                 .getResponse();
     }
