+++ tomcat/src/main/java/org/apache/coyote/model/response/HttpResponse.java
@@ -1,11 +1,21 @@
 package org.apache.coyote.model.response;
 
+import org.apache.coyote.exception.NotFoundFileException;
 import org.apache.coyote.model.session.Cookie;
 
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
+
+import static org.apache.coyote.model.request.ContentType.HTML;
 
 public class HttpResponse {
+
+    private static final String STATIC = "static";
 
     private final ResponseLine responseLine;
     private final ResponseHeader responseHeader;
@@ -27,6 +37,21 @@ public class HttpResponse {
         headers.put(ResponseHeader.CONTENT_TYPE, extension);
         headers.put(ResponseHeader.CONTENT_LENGTH, String.valueOf(body.getBytes().length));
         return ResponseHeader.of(headers);
+    }
+
+    public static String getResponseBody(final String uri, final Class<?> ClassType) {
+        try {
+            final URL url = Objects.requireNonNull(ClassType.getClassLoader().getResource(STATIC + uri));
+            final Path path = Paths.get(url.toURI());
+            return new String(Files.readAllBytes(path));
+        } catch (Exception e) {
+            throw new NotFoundFileException("파일 찾기에 실패했습니다.");
+        }
+    }
+
+    public static HttpResponse createResponse(final StatusCode statusCode, final String responseBody) {
+        ResponseLine responseLine = ResponseLine.of(statusCode);
+        return HttpResponse.of(HTML.getExtension(), responseBody, responseLine);
     }
 
     public void addCookie(final Cookie cookie) {
