+++ tomcat/src/main/java/org/apache/coyote/model/session/SessionManager.java
@@ -1,6 +1,5 @@
 package org.apache.coyote.model.session;
 
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
